refactor: table head main logic
